requirements.txt
# Core dependencies
streamlit==1.28.1
pandas==2.1.1
numpy==1.24.3
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
sqlite3

# Data ingestion
tweepy==4.14.0
praw==7.7.1
requests==2.31.0

# Sentiment analysis
vaderSentiment==3.3.2
transformers==4.34.0
torch==2.0.1

# Visualization
plotly==5.17.0
wordcloud==1.9.2
matplotlib==3.7.2
seaborn==0.12.2

# Pipeline orchestration
apache-airflow==2.7.2
prefect==2.13.4

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
retry==0.9.2

# Development
pytest==7.4.2
black==23.9.1
flake8==6.1.0